# test(bench): improve go-faster/jx benchmark