* Use ObjBytes to avoid key allocations
* Use Skip instead of parsing
* Avoid unrelated allocation (add strToBytes)

```
name                                     old time/op    new time/op    delta
CalcStats/gofaster-jx/tiny-32               237ns ± 2%      60ns ± 1%   -74.77%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/small-32             1.28µs ± 9%    0.88µs ± 4%   -30.85%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/large-32              151ms ± 1%      41ms ± 1%   -72.85%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/tiny-32      296ns ± 3%      74ns ± 2%   -75.08%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/small-32    2.10µs ± 6%    1.53µs ± 6%   -26.91%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/large-32     199ms ± 1%     164ms ± 2%   -17.66%  (p=0.008 n=5+5)

name                                     old alloc/op   new alloc/op   delta
CalcStats/gofaster-jx/tiny-32               40.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/small-32              80.0B ± 0%     16.0B ± 0%   -80.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/large-32             35.2MB ± 0%     0.0MB ± 7%  -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/tiny-32      40.0B ± 0%      0.0B       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/small-32      144B ± 0%       80B ± 0%   -44.44%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/large-32    57.9MB ± 0%    52.3MB ± 0%    -9.67%  (p=0.008 n=5+5)

name                                     old allocs/op  new allocs/op  delta
CalcStats/gofaster-jx/tiny-32                2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/small-32               8.00 ± 0%      2.00 ± 0%   -75.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx/large-32              1.12M ± 0%     0.00M       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/tiny-32       2.00 ± 0%      0.00       -100.00%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/small-32      17.0 ± 0%      13.0 ± 0%   -23.53%  (p=0.008 n=5+5)
CalcStats/gofaster-jx_withpath/large-32     1.77M ± 0%     1.54M ± 0%   -12.53%  (p=0.008 n=5+5)
```